+++ bin/test_coverage.dart
@@ -20,7 +20,14 @@ Future main(List<String> arguments) async {
         'Exclude specific files or directories using glob pattern (relative to package root), '
         'e.g. "subdir/*", "**_vm_test.dart".',
   );
-  parser.addFlag('badge', help: 'Generate coverage badge SVG image in your package root', defaultsTo: true);
+  parser.addOption('port',
+      abbr: 'p',
+      defaultsTo: '8787',
+      help: 'Set custom port for Dart Observatory to use when running tests.');
+
+  parser.addFlag('badge',
+      help: 'Generate coverage badge SVG image in your package root',
+      defaultsTo: true);
 
   final options = parser.parse(arguments);
 
@@ -34,12 +41,14 @@ Future main(List<String> arguments) async {
     excludeGlob = new Glob(options['exclude']);
   }
 
+  String port = options['port'];
+
   final testFiles = findTestFiles(packageRoot, excludeGlob: excludeGlob);
   print('Found ${testFiles.length} test files.');
   generateMainScript(packageRoot, testFiles);
   print('Generated test-all script in test/.test_coverage.dart. '
       'Please make sure it is added to .gitignore.');
-  await runTestsAndCollect(Directory.current.path).then((_) {
+  await runTestsAndCollect(Directory.current.path, port).then((_) {
     print('Coverage report saved to "coverage/lcov.info".');
   });
   final lcov = File(path.join(packageRoot.path, 'coverage', 'lcov.info'));